acoustic_bench/speech_recognition/pytorch/requirements.txt
nemo_toolkit
lightning
omegaconf
hydra-core
fiddle
cloudpickle
lhotse
braceexpand
editdistance
jiwer
kaldi-python-io
kaldialign
marshmallow
more-itertools
overrides
portalocker
webdataset
whisper_normalizer
pyannote.core
pyannote.database
pyannote.metrics
wandb
sacrebleu
sacremoses
inflect
pydub
pyloudnorm
sox
resampy
optuna
SQLAlchemy
colorlog
nv-one-logger-core
nv-one-logger-pytorch-lightning-integration
